% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_related_codes 2.R, R/cd_related_codes.R
\name{conv_to_cd_data}
\alias{conv_to_cd_data}
\title{Convert data frame as sparsebnData object}
\usage{
conv_to_cd_data(data)

conv_to_cd_data(data)
}
\arguments{
\item{data}{blah}
}
\value{
blah

blah
}
\description{
Convert data frame as sparsebnData object

Convert data frame as sparsebnData object
}
