% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eval 2.R, R/load_eval.R
\name{eval_by_lam}
\alias{eval_by_lam}
\title{Calculate Evaluation Metrics}
\usage{
eval_by_lam(A_est_by_lam, A_true, n_edges)

eval_by_lam(A_est_by_lam, A_true, n_edges)
}
\arguments{
\item{A_est_by_lam}{blah}

\item{A_true}{blah}

\item{n_edges}{blah}
}
\value{
blah

blah
}
\description{
Calculate Evaluation Metrics

Calculate Evaluation Metrics
}
